+++ CHANGELOG.md
@@ -8,11 +8,74 @@ If you like this project and find it useful, please consider giving it a star on
   <img src="bmc-button.svg" alt="Buy me a coffee" width="120">
 </a>
 
-## [3.2.10] - Not released
+## Project evolution
+
+The project will evolve to multi thread (the cli will become the threads manager) with these initial threads:
+
+- matterbridge;
+- frontend;
+- all plugins in bridge mode;
+- each plugin in childbridge mode;
+
+Advantages:
+
+- real concurrency outside the Node.js main loop;
+- isolation between threads;
+- single plugin isolation in childbridge mode;
+
+## [3.3.0] - Not released
 
 ### Development Breaking Changes
 
-- [platform]: Internal use only properties are private readonly.
+- [platform]: Now, internal use only properties are private readonly and internal use only methods are private. This prepares the road to plugin isolation.
+- [platform]: The signature of the matterbridge param in the platform constructor has changed from Matterbridge to `PlatformMatterbridge` that has only the approriate readonly properties from matterbridge. This prepares the road to plugin isolation.
+
+This change will require to adapt all plugins in two steps.
+
+1. `After` matterbridge `3.3.0` will be published as latest:
+
+- update the plugin platform constructor with the new signature:
+
+```typescript
+  constructor(matterbridge: PlatformMatterbridge, log: AnsiLogger, config: PlatformConfig)
+```
+
+- require matterbridge 3.3.0:
+
+```typescript
+    if (this.verifyMatterbridgeVersion === undefined || typeof this.verifyMatterbridgeVersion !== 'function' || !this.verifyMatterbridgeVersion('3.3.0')) {
+      throw new Error(
+        `This plugin requires Matterbridge version >= "3.3.0". Please update Matterbridge from ${this.matterbridge.matterbridgeVersion} to the latest version."`,
+      );
+    }
+```
+
+- check that you didn't used any matterbridge calls.
+
+In this phase (matterbridge 3.3.x) all plugins will continue to build and run even without updates.
+
+2. `After` matterbridge `3.4.0` will be published as latest, the new signature `PlatformMatterbridge` with the plugin isolation will be effective.
+
+```typescript
+export type PlatformMatterbridge = {
+  readonly homeDirectory: string;
+  readonly rootDirectory: string;
+  readonly matterbridgeDirectory: string;
+  readonly matterbridgePluginDirectory: string;
+  readonly globalModulesDirectory: string;
+  readonly matterbridgeVersion: string;
+  readonly matterbridgeLatestVersion: string;
+  readonly matterbridgeDevVersion: string;
+  readonly bridgeMode: 'bridge' | 'childbridge' | 'controller' | '';
+  readonly restartMode: 'service' | 'docker' | '';
+  readonly aggregatorVendorId: VendorId;
+  readonly aggregatorVendorName: string;
+  readonly aggregatorProductId: number;
+  readonly aggregatorProductName: string;
+};
+```
+
+In this phase (matterbridge 3.4.x) all plugins will not build and will not run without updates.
 
 ### Added